+++ exonum/src/node/mod.rs
@@ -86,8 +86,6 @@ pub enum ExternalMessage {
     Enable(bool),
     /// Shutdown the node.
     Shutdown,
-    /// Rebroadcast transactions from the pool.
-    Rebroadcast,
 }
 
 /// Node timeout types.